rename the component
